client/requirements.txt
mss
Pillow
websocket-client
